Fix bug ("saved draft" toast not shown despite draft being saved)
When the user moves from the editor to the inmate search screen, the draft is auto-saved since we need to be able to load it in when the user navigates back to the editor. Hence, we should show the "saved draft" toast if the user navigates back to the home screen when the draft already has a selected inmate.